requirements.txt
Python 3.12.3
Streamlit
plotly.express
